+++ src/components/Auth/Login/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { login } from '../actions';
-import Form from './Form';
+import Form from '../Form';
 export interface IAppProps {
   login: any
 }
@@ -10,7 +10,8 @@ const Login = (props:IAppProps) =>  {
 
   return (
     <>
-      <Form login={props.login}/>
+      <h1>Login</h1>
+      <Form onSubmit={props.login}/>
     </>
   );
 }
